runtime: el reproductor dibuja de verdad, y forge-escena deja de estar duplicada
`forge-runtime` abria cualquier documento, dibujaba una imagen vacia y decia que
todo habia ido bien. Tres fallos encadenados, cada uno silencioso:

  1. `cargar_geometria` sacaba los bytes del blob, los tiraba (`_malla_bytes`) y
     construia una `CpuMesh` vacia con un comentario que decia "por ahora". La
     malla vacia no pasaba `es_valida()` y se descartaba sin ruido.
  2. `renderizar` construia `instances = Vec::new()` a proposito.
  3. `main` cargaba los blobs del archivo en un `MemoryBlobStore` local dentro
     de `cargar_documento`, lo dejaba morir al volver, y luego renderizaba
     contra otro almacen recien creado y vacio.

Y los dos tests que habia usaban un documento **vacio**, con lo cual pasaban.

Habia ademas un desajuste de contrato entre dos piezas escritas por separado que
no se ve leyendo ninguna de las dos: `DrawInstance::mesh` lleva el hash del blob
**del documento**, mientras que `MapaDeMallas::insertar` indexa por el hash del
**contenido decodificado**. No son el mismo numero. Montar una escena con esa
pareja no da ningun error: `malla()` devuelve `None`, la instancia "se cuenta
como descartada y el frame sigue", y sale negro. Se anade
`insertar_con_hash(h, m)` para el camino del documento y se explica la trampa en
los tres sitios donde alguien se la puede encontrar.

**El formato del blob de malla, que no estaba decidido.** El documento
referencia `{"Mesh": hash}` y la especificacion no decia que hay dentro. Ahora
si (docs/formato/README.md, 4.1): un GLB escrito con `GltfOptions::crudo()`, o
sea milimetros y Z arriba, las unidades de FORGE. Incumple glTF a proposito y
por eso es normativo decirlo: leerlo con las opciones por defecto da los ejes
permutados y la escala mil veces mayor, sin error. Se elige GLB porque ya hay
lector y escritor en `forge-interop` -- una sola ruta de codigo para el
documento y para la exportacion -- y porque un blob suelto se puede inspeccionar
con herramientas que ya existen.

**`forge-escena`, caja nueva.** La traduccion `Snapshot` -> `DrawInstance` vivia
en `forge-ui`, y su propia documentacion decia que servia para que "un futuro
`forge-runtime` comparta exactamente este mismo camino". Cuando ese runtime
llego no podia: depender de `forge-ui` es justo lo que un reproductor sin editor
no puede hacer. Copiar la funcion habria dado dos versiones de la regla de que
se dibuja, que se separan a la primera. Se saca a su propia caja, `forge-ui` la
reexporta y su API publica no cambia. El guardia de fronteras lleva la entrada
nueva y la arista `forge-runtime -> forge-interop`, con el porque escrito.

Un blob que falta o que no es un GLB ahora es un **error** que dice cual, no un
objeto que se salta: una escena a medias sin avisar es peor que no abrirla.

Ocho tests de integracion contra geometria de verdad, entre ellos el que separa
"el reproductor funciona" de lo que habia: la escena con geometria tiene que dar
una imagen distinta de la de una escena vacia, y en mas del 1 % de los pixeles.
Mas la caja de mundo exacta en milimetros -- que falla si el blob se leyera con
las opciones de glTF -- el recuento de triangulos, tres instancias de un mismo
blob decodificado una sola vez, y un control de que las dos lecturas del GLB de
verdad difieren.

Ejemplo `escena_demo`: escribe un `.forge` con tres cajas, lo relee del disco
con un almacen limpio y lo dibuja. Comprobado de punta a punta, tambien con el
binario: 4 entidades, 3 instancias, 36 triangulos, caja (-160,-40,0)..(160,40,80).

De paso, dos luces en vez de una. Con una sola direccional y sin entorno las
caras que no miran a la luz salen negras sobre fondo negro, y las cajas parecian
huecas cuando la geometria estaba bien.

Tambien: `calcular_estadisticas` devolvia siempre cero triangulos y una caja
`None` (tenia un `let _bbox: Option<Aabb> = None;` que se asignaba al final sin
haberse tocado). Ahora cuenta y encuadra de verdad, y `renderizar` encuadra la
camara sobre la caja de la escena -- sin eso, cualquier escena que no este en el
origen sale fuera de cuadro.

Y las advertencias de clippy que quedaban en el codigo rescatado.

+++ crates/forge-render/src/camara.rs
@@ -377,7 +377,7 @@ mod tests {
             // redondeo de ese paso es esperable y no lo que se está probando.
             for i in 0..4 {
                 assert!(
-                    (obtenido[i] - esperado[i as usize]).abs() < 1e-2,
+                    (obtenido[i] - esperado[i]).abs() < 1e-2,
                     "componente {i}: obtenido {obtenido:?} vs esperado {esperado:?}"
                 );
             }
